Move delete account functionality into its own component
