+++ src/components/Buttons/index.ts
@@ -1,1 +1,4 @@
+import DeleteAccountButton from "./DeleteAccountButton"
+
 export { default as LinkButton } from "./LinkButton"
+export { default as DeleteAccountButton } from "./DeleteAccountButton"
